Home task 2, each of comment and comment-list components is separated to own folder
