+++ src/components/comment-list/style.css
@@ -1,0 +1,26 @@
+.comments-enter {
+    opacity: 0.01;
+}
+
+.comments-enter.comments-enter-active {
+    opacity: 1;
+    transition: opacity 500ms ease-in;
+}
+
+.comments-appear {
+    opacity: 0.01;
+}
+
+.comments-appear.comments-appear-active {
+    opacity: 1;
+    transition: opacity 1000ms ease-in;
+}
+
+.comments-leave {
+    opacity: 1;
+}
+
+.comments-leave.comments-leave-active {
+    opacity: 0.01;
+    transition: opacity 300ms ease-in;
+}
